refactor: DurationUnit enum化、Duration.note必須化、Step.position型縮小
PR #18 レビュー対応 Phase 1: モデル層の改修

- DurationValue.unit を String から DurationUnit enum に変更
  - 有効な値: Day, Hour, Minute, Second
  - 型レベルで不正な単位を防止
- Duration.note を Option<String> から String に変更（必須化）
- Step.position を i32 から i16 に変更（100ステップ以上は想定外）
- ParameterContent に JSON 構造例のドキュメントコメント追加
- GraphQL層でDurationUnit変換処理を追加
- EmptyUnit バリデーションエラーを削除（enum化で不要に）

+++ backend/src/domain/models/step.rs
@@ -29,7 +29,7 @@ pub struct Step {
     id: StepId,
     trial_id: TrialId,
     name: String,
-    position: i32,
+    position: i16,
     started_at: Option<chrono::DateTime<chrono::Utc>>,
     completed_at: Option<chrono::DateTime<chrono::Utc>>,
     parameters: Vec<Parameter>,
@@ -39,7 +39,7 @@ pub struct Step {
 
 impl Step {
     /// 新しいステップを作成する（ID は自動生成）
-    pub fn new(trial_id: TrialId, name: String, position: i32) -> Self {
+    pub fn new(trial_id: TrialId, name: String, position: i16) -> Self {
         let now = chrono::Utc::now();
         Self {
             id: StepId::new(),
@@ -60,7 +60,7 @@ impl Step {
         id: StepId,
         trial_id: TrialId,
         name: String,
-        position: i32,
+        position: i16,
         started_at: Option<chrono::DateTime<chrono::Utc>>,
         completed_at: Option<chrono::DateTime<chrono::Utc>>,
         parameters: Vec<Parameter>,
@@ -92,7 +92,7 @@ impl Step {
         &self.name
     }
 
-    pub fn position(&self) -> i32 {
+    pub fn position(&self) -> i16 {
         self.position
     }
